requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http
apache-airflow-providers-common-io
apache-airflow-providers-amazon[s3fs]
apache-airflow-providers-cncf-kubernetes
astronomer-providers[all]
astronomer-starship
astronomer-cosmos
# Libraries
polars
sqlalchemy
dbt-core==1.8.7
dbt-postgres==1.8.2
pandas
memory-profiler
connectorx
adbc-driver-snowflake